rename /users/router <- /users/user-router
