+++ server.js
@@ -1,13 +1,11 @@
 const express = require('express');
 const helmet = require('helmet');
 
-const UserRouter = require('./users/user-router.js');
-
 const server = express();
 
 server.use(helmet());
 server.use(express.json());
 
-server.use('/api/users', UserRouter);
+server.use('/api/users', require('./users/router'));
 
 module.exports = server;
